==> ECU/2018/2.D6m/Rcodes/5.Validaciones.R <==
#########################################################
# Proyecto MRP - Leave no one behind                    #
# Mapas de pobreza CEPAL                                #
# Lectura y preparación de las bases de datos           #
# Autor: Stalyn Guerrero & Andrés Gutiérrez             #
#########################################################

rm(list =ls())

# Librerías

library(tidyverse)
library(patchwork)
library(survey)
library(srvyr)

source("0Funciones/funciones_mrp.R", encoding = "UTF-8")
# Loading data ------------------------------------------------------------
encuesta_mrp <- readRDS("ECU/2018/2.D6m/Data/encuesta_mrp.rds")
poststrat_df <- readRDS("ECU/2018/2.D6m/Data/poststrat_df.RDS")

# Revisión de NAs ---------------------------------------------------------

View(poststrat_df)
sum(complete.cases(poststrat_df)) == nrow(poststrat_df)

# Calculo de la pobreza Cepal.  -------------------------------------------

poststrat_df %<>% mutate(yk_lmer = pobreza, 
                         yk_bench = pobreza2)

diseno <- encuesta_mrp %>%
  mutate(yk_dir = pobreza) %>% 
  as_survey_design(weights = fexp)

## validación nacional.
diseno %>% summarise(Nacional_dir = survey_mean(yk_dir))
poststrat_df %>% summarise(
  Nacional_lmer = sum(n * yk_lmer)/sum(n) ,
  Nacional_bench = sum(n * yk_bench)/sum(n)
)


###########################################
###########################################
### Validaciones por subgrupo completo  ###
###########################################
###########################################
bynames <-
  grep(
    pattern =  "^(X|F|n|pobreza|ingreso|tasa_desocupacion|epred_mat|gk|yk|urban.coverfraction|stable_lights|crops.coverfraction|sexo|mpio)",
    x = names(poststrat_df),
    invert = TRUE,
    value = TRUE
  )

bynames

plot_uni <- map(
  .x = setNames(bynames, bynames),
  ~ plot_compare2(
    sample_diseno = diseno,
    poststrat = poststrat_df,
    by1 = .x
  )
)

plot_uni

plot_uni <- (plot_uni$area$Plot$plot1 + plot_uni$anoest$Plot$plot1)/
      (plot_uni$edad$Plot$plot1+ plot_uni$discapacidad$Plot$plot1+ plot_uni$etnia$Plot$plot1)

ggsave(plot = plot_uni, 
       filename = "ECU/2018/2.D6m/Output/plot_uni.jpeg",scale = 5)


plot_uni


###########################################
###########################################
### Validaciones por pares de subgrupos ###
###########################################
###########################################

# # 1. Dpto-sexo -----------------------------------------------------------------

# bynames2 <- t(combn(bynames, 2))

# map(
#   1:nrow(bynames2),
#   ~ plot_compare2(
#     sample_diseno = diseno,
#     poststrat = poststrat_df,
#     by1 = bynames2[.x, ]
#   )$Plot$plot1 %>%
#     ggsave(
#       paste0(
#         "ECU/2018/2.D6m/Output/plot_",
#         paste0(bynames2[.x, ], collapse = "_"),
#         ".jpeg"
#       ),
#       plot = .,
#       scale = 3
#     )
# )



# ###########################################
# ###########################################
# ### Validaciones por triplas            ###
# ###########################################
# ###########################################
# bynames3 <- t(combn(bynames, 3))
# # 1. Dpto -----------------------------------------------------------------
# map(
#   1:nrow(bynames3),
#   ~ plot_compare2(
#     sample_diseno = diseno,
#     poststrat = poststrat_df,
#     by1 = bynames3[.x, ]
#   )$Plot$plot1 %>%
#     ggsave(
#       paste0(
#         "ECU/2018/2.D6m/Output/plot_",
#         paste0(bynames3[.x, ], collapse = "_"),
#         ".jpeg"
#       ),
#       plot = .,
#       scale = 4
#     )
# )
